chore: 네이버 핵 데이 컨벤션 적용
