+++ src/main/java/org/c4marathon/assignment/board/dto/BoardGetAllResponse.java
@@ -3,9 +3,9 @@ package org.c4marathon.assignment.board.dto;
 import java.time.LocalDateTime;
 
 public record BoardGetAllResponse(
-        String title,
-        String writerName,
-        LocalDateTime createdDate,
-        LocalDateTime lastModifiedDate
+	String title,
+	String writerName,
+	LocalDateTime createdDate,
+	LocalDateTime lastModifiedDate
 ) {
 }
