+++ src/main/java/org/c4marathon/assignment/board/dto/BoardUpdateResponse.java
@@ -3,11 +3,11 @@ package org.c4marathon.assignment.board.dto;
 import java.time.LocalDateTime;
 
 public record BoardUpdateResponse(
-        Long id,
-        String content,
-        String title,
-        String writerName,
-        LocalDateTime createdDate,
-        LocalDateTime lastModifiedDate
+	Long id,
+	String content,
+	String title,
+	String writerName,
+	LocalDateTime createdDate,
+	LocalDateTime lastModifiedDate
 ) {
 }
