+++ src/main/java/org/c4marathon/assignment/board/exception/NotFoundBoardException.java
@@ -1,11 +1,11 @@
 package org.c4marathon.assignment.board.exception;
 
+import static org.c4marathon.assignment.global.exception.ErrorCode.*;
+
 import org.c4marathon.assignment.global.exception.ApplicationException;
 
-import static org.c4marathon.assignment.global.exception.ErrorCode.NOT_FOUND_BOARD_ERROR;
-
 public class NotFoundBoardException extends ApplicationException {
-    public NotFoundBoardException() {
-        super(NOT_FOUND_BOARD_ERROR);
-    }
+	public NotFoundBoardException() {
+		super(NOT_FOUND_BOARD_ERROR);
+	}
 }
