--- src/main/java/org/c4marathon/assignment/user/domain/service/UserGetService.java
@@ -1,33 +1,0 @@
-package org.c4marathon.assignment.user.domain.service;
-
-import lombok.RequiredArgsConstructor;
-import org.c4marathon.assignment.global.annotation.DomainService;
-import org.c4marathon.assignment.user.domain.User;
-import org.c4marathon.assignment.user.domain.repository.UserRepository;
-import org.c4marathon.assignment.user.exception.NotFoundUserException;
-
-@DomainService
-@RequiredArgsConstructor
-public class UserGetService {
-
-    private final UserRepository userRepository;
-
-    public User getByEmail(String email) {
-        return userRepository.findByEmail(email)
-                .orElseThrow(() -> new NotFoundUserException());
-    }
-
-    public User getById(Long id) {
-        return userRepository.findById(id)
-                .orElseThrow(() -> new NotFoundUserException());
-    }
-
-    public boolean existByEmail(String email) {
-        return userRepository.existsByEmail(email);
-    }
-
-    public boolean existByNickname(String nickname) {
-        return userRepository.existsByNickname(nickname);
-    }
-
-}
