--- src/main/java/org/c4marathon/assignment/user/domain/service/UserSaveService.java
@@ -1,17 +1,0 @@
-package org.c4marathon.assignment.user.domain.service;
-
-import lombok.RequiredArgsConstructor;
-import org.c4marathon.assignment.global.annotation.DomainService;
-import org.c4marathon.assignment.user.domain.User;
-import org.c4marathon.assignment.user.domain.repository.UserRepository;
-
-@DomainService
-@RequiredArgsConstructor
-public class UserSaveService {
-
-    private final UserRepository userRepository;
-
-    public User save(User user) {
-        return userRepository.save(user);
-    }
-}
