+++ src/main/java/org/c4marathon/assignment/user/exception/DuplicatedNicknameException.java
@@ -1,11 +1,11 @@
 package org.c4marathon.assignment.user.exception;
 
+import static org.c4marathon.assignment.global.exception.ErrorCode.*;
+
 import org.c4marathon.assignment.global.exception.ApplicationException;
 
-import static org.c4marathon.assignment.global.exception.ErrorCode.NICKNAME_DUPLICATED_ERROR;
-
 public class DuplicatedNicknameException extends ApplicationException {
-    public DuplicatedNicknameException() {
-        super(NICKNAME_DUPLICATED_ERROR);
-    }
+	public DuplicatedNicknameException() {
+		super(NICKNAME_DUPLICATED_ERROR);
+	}
 }
