+++ src/main/java/org/c4marathon/assignment/user/exception/NotFoundUserException.java
@@ -1,11 +1,11 @@
 package org.c4marathon.assignment.user.exception;
 
+import static org.c4marathon.assignment.global.exception.ErrorCode.*;
+
 import org.c4marathon.assignment.global.exception.ApplicationException;
 
-import static org.c4marathon.assignment.global.exception.ErrorCode.NOT_FOUND_USER_ERROR;
-
 public class NotFoundUserException extends ApplicationException {
-    public NotFoundUserException() {
-        super(NOT_FOUND_USER_ERROR);
-    }
+	public NotFoundUserException() {
+		super(NOT_FOUND_USER_ERROR);
+	}
 }
